crud done

+++ backend/controllers/goalController.js
@@ -2,7 +2,7 @@ const asyncHandler = require("express-async-handler");
 
 
 const Goal = require('../models/goalModel');
-const userModel = require("../models/userModel");
+const User = require("../models/userModel");
 
 
 
@@ -37,6 +37,9 @@ const setGoals = asyncHandler( async(req, res) => {
 //@desc Update goals from db
 //@ route PUT /api/goals
 //@access Private
+
+
+
 const updateGoals =asyncHandler( async(req, res) => {
 
     const goal= await Goal.findById(req.params.id); //passing the id
@@ -44,6 +47,19 @@ const updateGoals =asyncHandler( async(req, res) => {
     if(!goal){
         res.status(404)
         throw new Error('Goal not found');
+
+
+        }
+
+        const user=await User.findById(req.user.id); //passing the id
+        if(!user){
+            res.status(404)
+            throw new Error('User not found');
+        }
+//make sure user owns the goal, logged in user is  the owner of the goal
+        if(goal.user.toString() !== user.id){
+            res.status(401)
+            throw new Error('User not authorized');
         }
 
         const updatedGoal= await Goal.findByIdAndUpdate(req.params.id, req.body, { //id of the goal, body of the request
@@ -64,6 +80,15 @@ const deleteGoals = asyncHandler(async(req, res) => {
     if(!goal){
         res.status(404)
         throw new Error('Goal not found');
+        }
+        const user=await User.findById(req.user.id); //passing the id
+        if(!user){
+            res.status(404)
+            throw new Error('User not found');
+        }
+//make sure user owns the goal, logged in user is  the owner of the goal
+        if(goal.user.toString() !== user.id){
+            res.status(401)
         }
 
         await Goal.findByIdAndDelete(req.params.id);
